the 'key_name' variable in the ec2 module in use here is looking for an aws keypair name (which always hold public key-halves), not the contents of an ssh private key!
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-key-pairs.html

fix is to create a keypair with the aws_key_pair resource and pass the key_name to the aws_instance resource. if this is done inside the module, we can remove one variable from the module's API, simplifying its use for other people who might need an ec2 module.

i've removed the ssh.pem files as they are not needed here...